+++ api/x-konnect/v1alpha1/schema_types.go
@@ -458,12 +458,6 @@ type DcrProviderDisplayName string
 // This is used to identify the DCR provider in the Konnect UI.
 type DcrProviderName string
 
-// GatewayDescription A human-readable description of the Gateway.
-type GatewayDescription string
-
-// GatewayName The name of the Gateway.
-type GatewayName string
-
 // LabelsValue is the value type for Labels.
 //
 // +kubebuilder:validation:MinLength=1
@@ -494,9 +488,14 @@ type LabelsUpdateValue string
 // "konnect", "mesh", "kic", or "_".
 type LabelsUpdate map[string]LabelsUpdateValue
 
-// MinRuntimeVersion The minimum runtime version supported by the API.
-// This is the lowest version of the data plane
-// release that can be used with the entity model.
-// When not specified, the minimum runtime version will be pinned to the latest
-// available release.
-type MinRuntimeVersion string
+// SourceIPEnabled Whether ip allow list is enabled for the organization.
+//
+// +kubebuilder:validation:Enum=Enabled;Disabled
+type SourceIPEnabled string
+
+const (
+	// SourceIPEnabledEnabled sets SourceIPEnabled as enabled.
+	SourceIPEnabledEnabled  SourceIPEnabled = "Enabled"
+	// SourceIPEnabledDisabled sets SourceIPEnabled as disabled.
+	SourceIPEnabledDisabled SourceIPEnabled = "Disabled"
+)